Create Assignment 2 LIS 4370
First section was given and second second was changes I made to get code to run correctly and provider the mean. The first section would not run correctly because the object assignment was not found it needed to be changed to assignment2. Also object someData was not found so that needed to be changed to assignment2 as well. Lastly I needed to add the call of myMean(assignment2) to get the mean to print out correctly. Without calling it then it will not appear.